services/image-ranker/requirements.txt
confluent-kafka
requests
Pillow
numpy
opencv-python-headless
boto3
python-dotenv
